Setup the CoreCLR repo to use the LocalPackagePublish mechanism.
This will enable the 'local publish' of the CoreCLR package to a location where it can be picked up by the CoreFx build.
What this means is that you will be able to build the CoreCLR package and use that to run the CoreFx tests on top of it.

Changes required to make this happen:
 - Take a dependency to the Microsoft.DotNet.BuildTools to bring in the packages.targets
 - Update the $(PackagesDir) to point to the <repo root>\packages to make it consistent with CoreFx
 - Hook-up the LocalPackagePublish targets to the build of the CoreCLR NuGet package